code cleanup

+++ BootEngine/Platforms/Windows/WindowsWindow.cs
@@ -80,7 +80,7 @@ namespace Platforms.Windows
 			//#TODO treat it here or hope for Veldrid to be updated with this (Maybe clone and alter source code myself)
 			window.Resized += () =>
 			{
-				EventCallback(new WindowResizeEvent((uint)window.Width, (uint)window.Height));
+				EventCallback(new WindowResizeEvent(window.Width, window.Height));
 			};
 
 			window.Closed += () => EventCallback(new WindowCloseEvent());
